Remove ContractChecker
