+++ src/main/java/randoop/test/ContractCheckingVisitor.java
@@ -8,6 +8,13 @@ import randoop.contract.ObjectContract;
 import randoop.sequence.ExecutableSequence;
 import randoop.sequence.ReferenceValue;
 import randoop.test.predicate.ExceptionPredicate;
+import randoop.types.ClassOrInterfaceType;
+import randoop.types.GenericClassType;
+import randoop.types.InstantiatedType;
+import randoop.types.ReferenceType;
+import randoop.types.Substitution;
+import randoop.types.Type;
+import randoop.types.TypeTuple;
 import randoop.util.Log;
 import randoop.util.TupleSet;
 
@@ -133,14 +140,11 @@ public final class ContractCheckingVisitor implements TestCheckGenerator {
                 + tuple.size()
                 + " must match contract arity "
                 + contract.getArity();
-        // if (Randomness.selectionLog.enabled() && Randomness.verbosity > 0) {
-        //   Randomness.selectionLog.log("ContractChecker.apply: considering contract=%s%n", contract);
-        // }
-        if (ContractChecker.typesMatch(contract.getInputTypes(), tuple)) {
+        if (typesMatch(contract.getInputTypes(), tuple)) {
           if (Log.isLoggingOn()) {
             Log.logLine("Checking contract " + contract.getClass());
           }
-          Object[] values = ContractChecker.getValues(tuple);
+          Object[] values = getValues(tuple);
           Check check = contract.checkContract(eseq, values);
           if (check != null) {
             return check;
@@ -150,5 +154,63 @@ public final class ContractCheckingVisitor implements TestCheckGenerator {
     }
 
     return null;
+  }
+
+  /**
+   * Indicates whether the given list of values matches the types in the type tuple. Contracts may
+   * have generic input types, so this method checks for consistent substitutions across value
+   * types.
+   *
+   * @param inputTypes the expected types for contract input
+   * @param valueTuple the values to match against input types
+   * @return true if the types of the values are assignable to the expected types, false otherwise
+   */
+  public static boolean typesMatch(TypeTuple inputTypes, List<ReferenceValue> valueTuple) {
+    if (inputTypes.size() != valueTuple.size()) {
+      return false;
+    }
+
+    Substitution<ReferenceType> substitution = new Substitution<>();
+    int i = 0;
+    while (i < inputTypes.size()) {
+      Type inputType = inputTypes.get(i);
+      ReferenceType valueType = valueTuple.get(i).getType();
+      if (inputType.isGeneric()) { // check substitutions
+        if (valueType instanceof ClassOrInterfaceType) {
+          ClassOrInterfaceType classType = (ClassOrInterfaceType) valueType;
+          InstantiatedType superType =
+              classType.getMatchingSupertype((GenericClassType) inputTypes.get(i));
+          if (superType == null) {
+            return false;
+          }
+          Substitution<ReferenceType> subst = superType.getTypeSubstitution();
+          if (!substitution.isConsistentWith(subst)) {
+            return false;
+          }
+          substitution = substitution.extend(subst);
+        } else { // have generic input type, and non-class value
+          return false;
+        }
+      } else if (!inputType.isAssignableFrom(valueType)) {
+        return false;
+      }
+      i++;
+    }
+    return true;
+  }
+
+  // TODO: Is this called a lot of times redundantly?
+  /**
+   * Creates an {@code Object} array for the given value list.
+   *
+   * @param tuple the list of values
+   * @return the Object array for the values
+   */
+  public static Object[] getValues(List<ReferenceValue> tuple) {
+    Object[] values = new Object[tuple.size()];
+    for (int i = 0; i < tuple.size(); i++) {
+      values[i] = tuple.get(i).getObjectValue();
+    }
+    return values;
   }
 }
